[log] Use uint64_t rather than size_t for writing/reading data
On 32 bits, size_t is 4 bytes while it is 8 bytes on 64 bits. This makes
all log generated on a 32 bits OS unreadable on a 64 bits OS and
vice-versa.

This patch works around this issue by using explicitly uint64_t
everywhere we used size_t for reading/writing purpose.

+++ src/mc_rtc/Logger.cpp
@@ -36,9 +36,9 @@ struct LoggerImpl
   std::ofstream log_;
 
 protected:
-  inline void fwrite(char * data, size_t size)
+  inline void fwrite(char * data, uint64_t size)
   {
-    log_.write((char *)&size, sizeof(size_t));
+    log_.write((char *)&size, sizeof(uint64_t));
     log_.write(data, static_cast<int>(size));
   }
